// Semester_3_Angular_Personal/Customer/src/app/models/account.model.ts
export class Account{
    accountId: number;
    firstname: string;
    lastname: string;
    email: string;
    password: string;
    phoneNumber: string;
    gender: string;
    addressId: number;
    avatar: string;
    roleId: number;
    status: boolean;
    securityCode: string;
    createdAt: string;
    updatedAt: string;
}